requirements.txt
langchain==0.2.16
langchain-core==0.2.43
langchain-google-genai==1.0.10
langgraph==0.2.0
google-generativeai==0.7.2
python-dotenv==1.0.1